:sparkles: main address validation

+++ src/main/java/com/github/sanaderer/POC/service/impl/AddressServiceImpl.java
@@ -5,6 +5,8 @@ import com.github.sanaderer.POC.controller.requests.AddressRequest;
 import com.github.sanaderer.POC.controller.responses.AddressResponse;
 import com.github.sanaderer.POC.entity.AddressEntity;
 import com.github.sanaderer.POC.entity.UserEntity;
+import com.github.sanaderer.POC.enums.Errors;
+import com.github.sanaderer.POC.exceptions.MainAddressException;
 import com.github.sanaderer.POC.repository.AddressRepository;
 import com.github.sanaderer.POC.service.AddressService;
 import com.github.sanaderer.POC.service.CepService;
@@ -33,21 +35,34 @@ public class AddressServiceImpl implements AddressService, CepService {
         return addressRepository.getReferenceById(id);
     }
 
-    public AddressEntity save(AddressRequest object, String cep, AddressEntity addressEntity) {
+    public AddressEntity save(AddressRequest addressRequest, String cep, AddressEntity addressEntity) {
         AddressResponse addressResponse = cepService.getCep(cep);
-        UserEntity userEntity = userService.findById(object.getUserId());
-        AddressEntity toEntity = mapper.toEntity(addressResponse, object, userEntity);
-        limitAddressesValidation(userEntity);
-        toEntity.setMainAddress(true);
+        UserEntity userEntity = userService.findById(addressRequest.getUserId());
+        AddressEntity toEntity = mapper.toEntity(addressResponse, addressRequest, userEntity);
+        limitAddressesValidation(userEntity, addressRequest);
+        toEntity.setMainAddress(addressRequest.getIsMain());
         return addressRepository.save(toEntity);
     }
 
-    private void limitAddressesValidation(UserEntity userEntity) {
+    private void limitAddressesValidation(UserEntity userEntity, AddressRequest addressRequest) {
         List<AddressEntity> list = addressRepository.findByUserOrderByDateCreated(userEntity);
-        list.forEach(addressEntity -> addressEntity.setMainAddress(false));
-        addressRepository.saveAllAndFlush(list);
-        if(list.size() >= 5){
-            addressRepository.delete(list.get(0));
+        if (list.isEmpty()) {
+            addressRequest.setIsMain(true);
+        } else {
+            if (list.size() >= 5) {
+                int index = Boolean.TRUE.equals(list.get(0).getMainAddress()) ? 1 : 0;
+                addressRepository.delete(list.get(index));
+            }
+        }
+    }
+
+    public void mainAddressesValidation(AddressEntity addressEntity, AddressRequest addressRequest) {
+        List<AddressEntity> addressEntities = addressRepository.findByUserOrderByDateCreated(addressEntity.getUser());
+        if (Boolean.TRUE.equals(addressRequest.getIsMain())) {
+            addressEntities.forEach(address -> address.setMainAddress(false));
+            addressRepository.saveAllAndFlush(addressEntities);
+        } else if (Boolean.TRUE.equals(addressEntity.getMainAddress())) {
+            throw new MainAddressException(Errors.PC204.getMessage(), Errors.PC204.getCode());
         }
     }
 
@@ -55,14 +70,17 @@ public class AddressServiceImpl implements AddressService, CepService {
         return cepService.getCep(cep);
     }
 
-
     public void deleteById(UUID id) {
+        if (Boolean.TRUE.equals(findById(id).getMainAddress())) {
+            throw new MainAddressException(Errors.PC204.getMessage(), Errors.PC204.getCode());
+        }
         addressRepository.deleteById(id);
     }
 
     public AddressEntity updateById(UUID id, AddressRequest object) {
-       AddressEntity addressEntity = addressRepository.findById(id).orElseThrow(RuntimeException::new);
-       return addressRepository.save(AddressMapper.toUpdateAddress(object, addressEntity));
+        AddressEntity addressEntity = addressRepository.findById(id).orElseThrow(RuntimeException::new);
+        mainAddressesValidation(addressEntity, object);
+        return addressRepository.save(AddressMapper.toUpdateAddress(object, addressEntity));
     }
 
 }
